Doc stat visualization (#415)
* document: stat_visualization

* document: stat_visualization

* document: stat_visualization

* update README

* versionup

* 不要なファイルを削除

* test code追加

* test code: project

* fix: job delete: ジョブIDの存在チェックを入れる

* update test

* update test: project_member

* update

* update

* update test

* bug: `annotation import`: annotation_idが含まれていないとインポートできない問題を修正

* update test

* format

* ``--qeury`` がゆうこうでない　場合を修正

* update document

* update document

* update document

* update README

+++ annofabcli/job/delete_job.py
@@ -23,12 +23,21 @@ class DeleteJobMain:
         self.facade = AnnofabApiFacade(service)
 
     def delete_job_list(self, project_id: str, job_type: JobType, job_id_list: List[str]):
+        job_list = self.service.wrapper.get_all_project_job(project_id, query_params={"type": job_type.value})
+        exists_job_id_set = {e["job_id"] for e in job_list}
+
+        count = 0
         for job_id in job_id_list:
-            logger.debug(f"job_id={job_id} のジョブを削除します。")
+            if job_id not in exists_job_id_set:
+                logger.debug(f"job_id={job_id} のジョブは存在しなかったのでスキップします。")
+                continue
             try:
                 self.service.api.delete_project_job(project_id, job_type.value, job_id)
+                logger.debug(f"job_type={job_type.value}, job_id={job_id} のジョブを削除しました。")
+                count += 1
             except Exception as e:  # pylint: disable=broad-except
                 logger.warning(e)
+        logger.info(f"{count} / {len(job_id_list)} 件のジョブを削除しました。")
 
 
 class DeleteJob(AbstractCommandLineInterface):
